+++ mods/BrewLAN/documentation/Changelog.md
@@ -207,6 +207,10 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   * R&D: Finished the UEF Penetration Fighter, Tomcat, graphically. It's missing weapons.
   * R&D: Added anti-missile flares to the penetration fighters.
 
+  ### 2018-05-11
+
+  * R&D: Finished the Tomcat.
+
 ## 2019-01-03|84 (FAF minor patch)
 
 * Created a function to add the FAF category CQUEMOV to function to mobile experimental mobile units.